Add exa alternative display format for "ls"
I've been hesitant to alias `ls` to `l` for quite some time thinking it
would be a crutch. However, the frequency with which I run it is
starting to weigh on me again. It would also be nice to have a nicer
default format to go along with it.

`exa` is one such tool that provides a prettier display for file listing
output. Let's try it out.