Add fd plugin
`fd` (https://github.com/sharkdp/fd) is a fast alternative to `find`.

In local tests in my home directory, `fd` was almost twice as fast as
the equivalent `ripgrep` command:

    time bash -c 'rg --files --hidden --follow > /dev/null &>/dev/null'
    0.99s user 2.68s system 595% cpu 0.615 total

    time bash -c "fd --hidden --type file > /dev/null &>/dev/null"
    0.60s user 1.55s system 642% cpu 0.335 total

`fd` uses some of the same dependencies as `ripgrep`, but is focused on
just finding files.

While here, swap out `fzf`'s default command to use `fd` for the speed.