Configure fzf to work with terminal key bindings
We didn't set this up when we first installed `fzf` since we didn't
understand the full feature set before diving in.

After a bit of reading, we're ready to activate.

One unfortunate workaround implemented was not using `brew --prefix` to
dynamically get the location of the shell completion/key-binding files.
The reason was this was terribly slow to run, and given we source these
files on every shell, we wanted it to remain performant.