Add ctags and default-gems plugins to rbenv
These two plugins install hooks that are automatically executed when a
new Ruby version is installed. With the settings I have, output looks
like:

    ⨠ rbenv install 2.5.3
    ruby-build: use openssl from homebrew
    Downloading ruby-2.5.3.tar.bz2...
    -> https://cache.ruby-lang.org/pub/ruby/2.5/ruby-2.5.3.tar.bz2
    Installing ruby-2.5.3...
    ruby-build: use readline from homebrew
    Installed ruby-2.5.3 to /Users/x/.rbenv/versions/2.5.3

    Running ctags on /Users/x/.rbenv/versions/2.5.3/lib/ruby/2.5.0
    Running ctags on /Users/x/.rbenv/versions/2.5.3/lib/ruby/site_ruby/2.5.0
    Running ctags on /Users/x/.rbenv/versions/2.5.3/lib/ruby/vendor_ruby/2.5.0
    Running ctags on /Users/x/.rbenv/versions/2.5.3/include/ruby-2.5.0
    Fetching: gem-ctags-1.0.9.gem (100%)
    Successfully installed gem-ctags-1.0.9
    1 gem installed
    Fetching: bundler-1.17.1.gem (100%)
    Generating ctags for bundler-1.17.1
    Injecting gem-ctags into bundler-1.17.1
    Successfully installed bundler-1.17.1
    1 gem installed

`rbenv-ctags` generates tags of the Ruby standard library after install.
`rbenv-default-gems` ensures some default gems are installed for all
Ruby versions (in this case, `gem-ctags` and `bundler`).

The `gem-ctags` gem ensures tags are generated automatically after each
gem is installed. It is first in the list so that all other gems in the
list have tags auto-generated after install.